# lib/api-zod/src/generated/types/exerciseProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym workout tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";
import type { ExerciseProgressPoint } from "./exerciseProgressPoint";

export interface ExerciseProgress {
  exercise: Exercise;
  points: ExerciseProgressPoint[];
}